+++ internal/authn/openid/adapter.go
@@ -1,4 +1,4 @@
-package oidc
+package openid
 
 import (
 	"log/slog"
