+++ internal/authn/openid/adapter_test.go
@@ -1,4 +1,4 @@
-package oidc
+package openid
 
 import (
 	"bytes"
