+++ internal/authn/openid/authn.go
@@ -1,4 +1,4 @@
-package oidc
+package openid
 
 import (
 	"context"
