+++ internal/authn/openid/authn_test.go
@@ -1,4 +1,4 @@
-package oidc
+package openid
 
 import (
 	"context"
