+++ internal/authn/openid/fakes_test.go
@@ -1,4 +1,4 @@
-package oidc
+package openid
 
 import (
 	"crypto/ecdsa"
